+++ pn-dcp/src/pn_dcp/ident_req.rs
@@ -34,7 +34,7 @@ impl BlockTrait for IdentReqBlocks {
         len
     }
 
-    fn payload(&self) -> usize {
+    fn payload(&self) -> u16 {
         unreachable!()
     }
 
@@ -53,7 +53,7 @@ impl BlockTrait for IdentReqBlock {
         }
     }
 
-    fn payload(&self) -> usize {
+    fn payload(&self) -> u16 {
         match self {
             Self::Block(a) => a.payload(),
             Self::Padding(a) => a.payload(),
